refactor: scrollbarview
