+++ src/components/playlist/SongList/SongList.tsx
@@ -4,6 +4,7 @@ import { FlashList } from '@shopify/flash-list';
 import { IconButton } from 'react-native-paper';
 import { useNetInfo } from '@react-native-community/netinfo';
 import { useFocusEffect } from '@react-navigation/native';
+import { useSharedValue } from 'react-native-reanimated';
 
 import { styles } from '@components/style';
 import SongInfo from './SongInfo';
@@ -13,7 +14,7 @@ import SongMenu from './SongMenu';
 import PlaylistInfo from '../Info/PlaylistInfo';
 import PlaylistMenuButton from '../Menu/PlaylistMenuButton';
 import usePlaylist from '../usePlaylistRN';
-import SongListScrollbar from './SongListScrollbar';
+import SongListScrollbar from './SongListScrollbar.new';
 import keepAwake from '@utils/keepAwake';
 
 export default () => {
@@ -40,9 +41,8 @@ export default () => {
     playlistRef,
   } = usedPlaylist;
   const netInfo = useNetInfo();
-  const [scrollViewHeight, setScrollViewHeight] = useState(0);
-  const [contentViewHeight, setContentViewHeight] = useState(0);
-  const [scrollPositionY, setScrollPositionY] = useState(0);
+  const scrollViewHeight = useSharedValue(0);
+  const scrollBarPosition = useSharedValue(0);
 
   useEffect(
     () => scrollTo({ toIndex: -1, reset: true }),
@@ -104,9 +104,8 @@ export default () => {
       <SongListScrollbar
         style={stylesLocal.playlistContainer}
         scrollViewReference={playlistRef}
+        scrollBarPosition={scrollBarPosition}
         scrollViewHeight={scrollViewHeight}
-        contentViewHeight={contentViewHeight}
-        scrollPositionY={scrollPositionY}
       >
         <FlashList
           ref={playlistRef}
@@ -134,24 +133,12 @@ export default () => {
           onRefresh={() => keepAwake(refreshPlaylist)}
           refreshing={refreshing}
           showsVerticalScrollIndicator={false}
-          onLayout={({
-            nativeEvent: {
-              layout: { height },
-            },
-          }) => {
-            setScrollViewHeight(height);
-          }}
-          onContentSizeChange={(_, contentHeight) => {
-            setContentViewHeight(contentHeight);
-          }}
           onScroll={({
             nativeEvent: { contentOffset, contentSize, layoutMeasurement },
           }) => {
-            setScrollPositionY(contentOffset.y);
-            setContentViewHeight(contentSize.height);
-            console.log(
-              contentOffset.y / (contentSize.height - layoutMeasurement.height),
-            );
+            scrollBarPosition.value =
+              contentOffset.y / (contentSize.height - layoutMeasurement.height);
+            scrollViewHeight.value = layoutMeasurement.height;
           }}
         />
         <SongMenu
